+++ database/migrations/2023_05_25_201041_create_expenses_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->dateTime('scheduled_at')->nullable();
             $table->dateTime('transacted_at')->nullable();
             $table->boolean('completed')->default(false);
-            $table->decimal('amount', 16)->default(0);
+            $table->bigInteger('amount')->default(0);
             $table->timestamps();
         });
     }
